+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes
 0.13.0 to 0.14.0
 ----------------
 * [UI]: Added ability to select or deselect viewing all associated projects. (0.13.1)
+* [UI]: Fixed issue with missing settings tab when in export view within a Project. (0.13.2)
+* [UI]: Fixed bug preventing sample selection while filtered by file. (0.13.2)
 
 0.12.0 to 0.13.0
 ----------------